Handle the case when the SVG does not have the xlink prefix. Closes #8
This was removed in SVG 2:
> SVG 2 removed the need for the xlink namespace, so instead of xlink:href you should use href.
> https://developer.mozilla.org/en-US/docs/Web/SVG/Attribute/xlink:href

.NET Framework supports this out the box with XmlReader and a Stream or string, but the other targets only support a Stream.
With .NET Standard, we have to use a Stream (thus we just File.OpenRead first). On PCL, we first try go via reflection an find the method that takes a string (we may be on .NET Framework), but if that fails, just use the old method that may possible fail on some systems. But, modern systems will use the .NET Standard library, so we should be safe.